+++ build-crystal-hell-locally.sh
@@ -1,0 +1,7 @@
+#!/bin/sh
+rm -rf Config
+mkdir Config
+go run .
+cp -r Config-Shared/* Config
+cp -r Config-CrystalHell/* Config
+