+++ data/aloe.go
@@ -2,33 +2,24 @@ package data
 
 import "fmt"
 
+// Aloe is a type of plant
 type Aloe struct {
 	Name               string
 	NamePlural         string
 	DisplayName        string
 	Description        string
-	PreferredConsumer  string
 	CropYield          int
 	BonusYield         int
 	CraftTime          int
 	incompatibleTraits []rune
 }
 
-func CreateAloe() *Aloe {
-	return &Aloe{
-		Name:              "Aloe",
-		DisplayName:       "Aloe Vera",
-		PreferredConsumer: "",
-		CropYield:         2,
-		BonusYield:        1,
-		CraftTime:         2,
-	}
-}
-
+// GetCraftTime returns the time required to craft this seed
 func (p *Aloe) GetCraftTime() int {
 	return p.CraftTime
 }
 
+// GetDescription returns the seed description for this plant
 func (p *Aloe) GetDescription() string {
 	if p.Description == "" {
 		return getDefaultSeedDescription()
@@ -36,22 +27,22 @@ func (p *Aloe) GetDescription() string {
 	return p.Description
 }
 
+// GetDisplayName returns the display name
 func (p *Aloe) GetDisplayName() string {
 	return p.DisplayName
 }
 
+// GetName returns the name of this plant
 func (p *Aloe) GetName() string {
 	return p.Name
 }
 
-func (p *Aloe) GetPreferredConsumer() string {
-	return p.PreferredConsumer
-}
-
+// GetSchematicName returns the schematic name for this plant, given the provided traits
 func (p *Aloe) GetSchematicName(traits string) string {
 	return fmt.Sprintf("plantedAloe1_%sschematic", traits)
 }
 
+// IsCompatibleWith checks for trait compatibility with this plant
 func (p *Aloe) IsCompatibleWith(t Trait) bool {
 	for _, incompatibleTrait := range p.incompatibleTraits {
 		if incompatibleTrait == t.Code {
@@ -61,13 +52,14 @@ func (p *Aloe) IsCompatibleWith(t Trait) bool {
 	return true
 }
 
+// WriteBlockStages produces each of the 3 block stages for this plant
 func (p *Aloe) WriteBlockStages(c chan string, target, traits string) {
-	p.WriteStage1(c, target, traits)
-	p.WriteStage2(c, traits)
-	p.WriteStage3(c, traits)
+	p.writeStage1(c, target, traits)
+	p.writeStage2(c, traits)
+	p.writeStage3(c, traits)
 }
 
-func (p *Aloe) WriteStage1(c chan string, target, traits string) {
+func (p *Aloe) writeStage1(c chan string, target, traits string) {
 	c <- fmt.Sprintf(`<block name="plantedAloe1_%s" stage="1" traits="%s">
 	<drop event="Destroy" name="plantedAloe1_%s" count="1"/>
 	<property name="CreativeMode" value="Player"/>
@@ -84,7 +76,7 @@ func (p *Aloe) WriteStage1(c chan string, target, traits string) {
 </block>`, traits, traits, traits, traits, getCraftingGroup(traits), traits, optionallyAddUnlock(p, target, traits))
 }
 
-func (*Aloe) WriteStage2(c chan string, traits string) {
+func (*Aloe) writeStage2(c chan string, traits string) {
 	c <- fmt.Sprintf(`<block name="plantedAloe2_%s" stage="2" traits="%s">
 	<property name="CreativeMode" value="Dev"/>
 	<property name="CustomIconTint" value="00ff80"/>
@@ -94,7 +86,7 @@ func (*Aloe) WriteStage2(c chan string, traits string) {
 </block>`, traits, traits, traits, traits)
 }
 
-func (p *Aloe) WriteStage3(c chan string, traits string) {
+func (p *Aloe) writeStage3(c chan string, traits string) {
 	c <- fmt.Sprintf(`<block name="plantedAloe3_%s" stage="3" traits="%s" tags="T%dPlant">
 	<drop event="Destroy" name="plantedAloe1_%s" count="1" prob="0.5"/>
 	<drop event="Fall" name="resourceYuccaFibers" count="0" prob="1" stick_chance="0"/>