+++ data/chrysanthemum.go
@@ -2,34 +2,24 @@ package data
 
 import "fmt"
 
+// Chrysanthemum is a type of plant
 type Chrysanthemum struct {
 	Name               string
 	NamePlural         string
 	DisplayName        string
 	Description        string
-	PreferredConsumer  string
 	CropYield          int
 	BonusYield         int
 	CraftTime          int
 	incompatibleTraits []rune
 }
 
-func CreateChrysanthemum() *Chrysanthemum {
-	return &Chrysanthemum{
-		Name:               "Chrysanthemum",
-		DisplayName:        "Chrysanthemum",
-		PreferredConsumer:  "Does",
-		CropYield:          2,
-		BonusYield:         1,
-		CraftTime:          2,
-		incompatibleTraits: []rune{},
-	}
-}
-
+// GetCraftTime returns the time required to craft this seed
 func (p *Chrysanthemum) GetCraftTime() int {
 	return p.CraftTime
 }
 
+// GetDescription returns the seed description for this plant
 func (p *Chrysanthemum) GetDescription() string {
 	if p.Description == "" {
 		return getDefaultSeedDescription()
@@ -37,22 +27,22 @@ func (p *Chrysanthemum) GetDescription() string {
 	return p.Description
 }
 
+// GetDisplayName returns the display name
 func (p *Chrysanthemum) GetDisplayName() string {
 	return p.DisplayName
 }
 
+// GetName returns the name of this plant
 func (p *Chrysanthemum) GetName() string {
 	return p.Name
 }
 
-func (p *Chrysanthemum) GetPreferredConsumer() string {
-	return p.PreferredConsumer
-}
-
+// GetSchematicName returns the schematic name for this plant, given the provided traits
 func (p *Chrysanthemum) GetSchematicName(traits string) string {
 	return fmt.Sprintf("plantedChrysanthemum1_%sschematic", traits)
 }
 
+// IsCompatibleWith checks for trait compatibility with this plant
 func (p *Chrysanthemum) IsCompatibleWith(t Trait) bool {
 	for _, incompatibleTrait := range p.incompatibleTraits {
 		if incompatibleTrait == t.Code {
@@ -62,13 +52,14 @@ func (p *Chrysanthemum) IsCompatibleWith(t Trait) bool {
 	return true
 }
 
+// WriteBlockStages produces each of the 3 block stages for this plant
 func (p *Chrysanthemum) WriteBlockStages(c chan string, target, traits string) {
-	p.WriteStage1(c, target, traits)
-	p.WriteStage2(c, traits)
-	p.WriteStage3(c, traits)
+	p.writeStage1(c, target, traits)
+	p.writeStage2(c, traits)
+	p.writeStage3(c, traits)
 }
 
-func (p *Chrysanthemum) WriteStage1(c chan string, target, traits string) {
+func (p *Chrysanthemum) writeStage1(c chan string, target, traits string) {
 	c <- fmt.Sprintf(`<block name="plantedChrysanthemum1_%s" stage="1" traits="%s">
 	<drop event="Destroy" name="plantedChrysanthemum1_%s" count="1"/>
 	<property name="CreativeMode" value="Player"/>
@@ -83,7 +74,7 @@ func (p *Chrysanthemum) WriteStage1(c chan string, target, traits string) {
 </block>`, traits, traits, traits, traits, getCraftingGroup(traits), traits, optionallyAddUnlock(p, target, traits))
 }
 
-func (*Chrysanthemum) WriteStage2(c chan string, traits string) {
+func (*Chrysanthemum) writeStage2(c chan string, traits string) {
 	c <- fmt.Sprintf(`<block name="plantedChrysanthemum2_%s" stage="2" traits="%s">
 	<property name="CreativeMode" value="Dev"/>
 	<property name="CustomIconTint" value="00ff80"/>
@@ -93,7 +84,7 @@ func (*Chrysanthemum) WriteStage2(c chan string, traits string) {
 </block>`, traits, traits, traits, traits)
 }
 
-func (p *Chrysanthemum) WriteStage3(c chan string, traits string) {
+func (p *Chrysanthemum) writeStage3(c chan string, traits string) {
 	c <- fmt.Sprintf(`<block name="plantedChrysanthemum3_%s" stage="3" traits="%s" tags="T%dPlant">
 	<drop event="Destroy" name="plantedChrysanthemum1_%s" count="1" prob="0.5"/>
 	<drop event="Fall" name="resourceYuccaFibers" count="0" prob="1" stick_chance="0"/>