+++ data/plant.go
@@ -5,31 +5,112 @@ import (
 	"strings"
 )
 
+// Plant interface describes the common interactions one can take with all plants
 type Plant interface {
 	GetCraftTime() int
 	GetDescription() string
 	GetDisplayName() string
 	GetName() string
-	GetPreferredConsumer() string
 	GetSchematicName(string) string
 	IsCompatibleWith(Trait) bool
 	WriteBlockStages(chan string, string, string)
 }
 
+// Plants is a collection of all plants to be processed
 var Plants []Plant = []Plant{
-	CreateAloe(),
-	CreateBlueberry(),
-	CreateChrysanthemum(),
-	CreateCoffee(),
-	CreateCorn(),
-	CreateCotton(),
-	CreateGoldenrod(),
-	CreateGraceCorn(),
-	CreateHop(),
-	CreateMushroom(),
-	CreatePotato(),
-	CreatePumpkin(),
-	CreateYucca(),
+	&Aloe{
+		Name:        "Aloe",
+		DisplayName: "Aloe Vera",
+		CropYield:   2,
+		BonusYield:  1,
+		CraftTime:   2,
+	},
+	&Blueberry{
+		Name:        "Blueberry",
+		DisplayName: "Blueberry",
+		CropYield:   2,
+		BonusYield:  1,
+		CraftTime:   2,
+	},
+	&Chrysanthemum{
+		Name:        "Chrysanthemum",
+		DisplayName: "Chrysanthemum",
+		CropYield:   2,
+		BonusYield:  1,
+		CraftTime:   2,
+	},
+	&Coffee{
+		Name:        "Coffee",
+		DisplayName: "Coffee",
+		CropYield:   2,
+		BonusYield:  1,
+		CraftTime:   2,
+	},
+	&Corn{
+		Name:        "Corn",
+		DisplayName: "Corn",
+		CropYield:   2,
+		BonusYield:  1,
+		CraftTime:   2,
+	},
+	&Cotton{
+		Name:        "Cotton",
+		DisplayName: "Cotton",
+		CropYield:   2,
+		BonusYield:  1,
+		CraftTime:   2,
+	},
+	&Goldenrod{
+		Name:        "Goldenrod",
+		DisplayName: "Goldenrod",
+		CropYield:   2,
+		BonusYield:  1,
+		CraftTime:   2,
+	},
+	&GraceCorn{
+		Name:        "GraceCorn",
+		DisplayName: "Super Corn",
+		CropYield:   2,
+		BonusYield:  1,
+		CraftTime:   2,
+	},
+	&Hop{
+		Name:        "Hop",
+		DisplayName: "Hop",
+		CropYield:   2,
+		BonusYield:  1,
+		CraftTime:   2,
+	},
+	&Mushroom{
+		Name:               "Mushroom",
+		DisplayName:        "Mushroom Spores",
+		Description:        `Mushroom spores can be planted on all surfaces and walls and will grow without sunlight.`,
+		CropYield:          2,
+		BonusYield:         1,
+		CraftTime:          2,
+		incompatibleTraits: []rune{'U'},
+	},
+	&Potato{
+		Name:        "Potato",
+		DisplayName: "Potato",
+		CropYield:   2,
+		BonusYield:  1,
+		CraftTime:   2,
+	},
+	&Pumpkin{
+		Name:        "Pumpkin",
+		DisplayName: "Pumpkin",
+		CropYield:   2,
+		BonusYield:  1,
+		CraftTime:   2,
+	},
+	&Yucca{
+		Name:        "Yucca",
+		DisplayName: "Yucca",
+		CropYield:   2,
+		BonusYield:  1,
+		CraftTime:   2,
+	},
 }
 
 func calculateCropYield(count int, traits string) int {
@@ -42,11 +123,6 @@ func calculateCropYield(count int, traits string) int {
 		count *= 4
 	} else if strings.Contains(traits, "B") {
 		count = int(float64(count) * 2)
-	}
-
-	// [R] Renewable
-	if strings.Contains(traits, "R") {
-		count = int(float64(count) * .75)
 	}
 
 	return count
@@ -83,5 +159,5 @@ func getDefaultSeedDescription() string {
 }
 
 func getCraftingGroup(traits string) string {
-	return fmt.Sprintf("Tier%dSeeds", len(traits) + 1)
+	return fmt.Sprintf("Tier%dSeeds", len(traits)+1)
 }