+++ CHANGELOG.md
@@ -5,10 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [TODO]
+## [0.6.0] - 2022-05-19
 
-- add seed description to schematic?
-- reconsider whether renewable trait should decrease yield
+- remove 25% decrease to crop yield for renewable
+- automate release build pipeline
+- add rotting flesh to enhanced seed recipe
 
 ## [0.5.0] - 2022-05-16